pyproject.toml
[project]
name = "kugl"
version = "0.6.0"
description = "Explore Kubernetes resources using SQLite"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Jon Ross", email = "kugl.devel@gmail.com"}
]
keywords = ["kubernetes", "kubectl", "sql", "sqlite"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "arrow>=1.0.1,<=1.3.0",
    "jmespath>=0.9.5,<=1.0",
    "pydantic>=2.0.2,<=2.9.2",
    "pyyaml>=5.3,<=6.0.2",
    "sqlparse>=0.5.1,<=0.5.3",
    "tabulate>=0.8.7,<=0.9.0",
]

[project.urls]
Homepage = "https://github.com/jonross/kugl"
Repository = "https://github.com/jonross/kugl"

[project.scripts]
kugl = "kugl.main:main"

[dependency-groups]
dev = [
    "pytest>=7.1.3",
    "pytest-cov>=4.1.0",
]
docs = [
    "sphinx-rtd-theme==3.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kugl"]

[tool.hatch.build.force-include]
"kugl/builtins/schemas/kubernetes.yaml" = "kugl/builtins/schemas/kubernetes.yaml"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-vv", "-s", "--tb=native", "--cov", "--cov-report=html:coverage"]

[tool.coverage.run]
omit = ["tests/kubectl/*", "tests/testing.py"]
